Fix identify regression on rasters
By setting ee.Image to visualize(**params) we were effectively creating a 3 band image and received the 3 bands from identify even if we were interested in the single band results for the elevation layer for example. The visualize(**params) is only passed to EarthEngine so it knows how to render the WMS layer.